+++ include/files.hpp
@@ -2,6 +2,7 @@
 #ifndef _FILE_FILES
 #define _FILE_FILES _FILE_FILES
 #include"init.hpp"
+#include"log.hpp"
 #include"settings.hpp"
 #define _run_ans 101
 #define _run_chk 102