+++ include/run.hpp
@@ -2,6 +2,7 @@
 #ifndef _FILE_RUN
 #define _FILE_RUN _FILE_RUN
 #include"init.hpp"
+#include"log.hpp"
 #include"settings.hpp"
 #include"files.hpp"
 #include"compile.hpp"