+++ include/settings.hpp
@@ -2,6 +2,7 @@
 #ifndef _FILE_SETTINGS
 #define _FILE_SETTINGS _FILE_SETTINGS
 #include"init.hpp"
+#include"log.hpp"
 json settings;
 namespace Settings
 {